Add DiscordService#sendMessage
